Module leptos_reactive::signal_prelude::prelude

This prelude imports all signal types as well as all signal traits needed to use those types.

Re-exports

pub use super::*;

Structs

Memo

An efficient derived reactive value based on other reactive values.

Signal

A wrapper for any kind of readable reactive signal: a ReadSignal, Memo, RwSignal, or derived signal closure.

SignalSetter

A wrapper for any kind of settable reactive signal: a WriteSignal, RwSignal, or closure that receives a value and sets a signal depending on it.

Enums

MaybeSignal

A wrapper for a value that is either T or Signal<T>.

Traits

IntoSignal

Helper trait for converting Fn() -> T closures into Signal<T>.

IntoSignalSetter

Helper trait for converting Fn(T) into SignalSetter<T>.

Functions

create_memo

Creates an efficient derived reactive value based on other reactive values.

create_selector

Creates a conditional signal that only notifies subscribers when a change in the source signal’s value changes whether it is equal to the key value (as determined by PartialEq.)

create_selector_with_fn

Creates a conditional signal that only notifies subscribers when a change in the source signal’s value changes whether the given function is true.