Struct leptos_reactive::signal_prelude::Signal

pub struct Signal<T>where
    T: 'static,{ /* private fields */ }

A wrapper for any kind of readable reactive signal: a ReadSignal, Memo, RwSignal, or derived signal closure.

This allows you to create APIs that take any kind of Signal<T> as an argument, rather than adding a generic F: Fn() -> T. Values can be access with the same function call, with(), and get() APIs as other signals.

Core Trait Implementations

.get() (or calling the signal as a function) clones the current value of the signal. If you call it within an effect, it will cause that effect to subscribe to the signal, and to re-run whenever the value of the signal changes.

.get_untracked() clones the value of the signal without reactively tracking it.

.with() allows you to reactively access the signalโs value without cloning by applying a callback function.

.with_untracked() allows you to access the signalโs value without reactively tracking it.

.to_stream() converts the signal to an async stream of values.

Examples

let (count, set_count) = create_signal(cx, 2);
let double_count = Signal::derive(cx, move || count() * 2);
let memoized_double_count = create_memo(cx, move |_| count() * 2);

// this function takes any kind of wrapped signal
fn above_3(arg: &Signal<i32>) -> bool {
    // โ calling the signal clones and returns the value
    //    it is a shorthand for arg.get()
    arg() > 3
}

assert_eq!(above_3(&count.into()), false);
assert_eq!(above_3(&double_count), true);
assert_eq!(above_3(&memoized_double_count.into()), true);

Implementations

impl<T> Signal<T>where T: 'static,

pub fn derive(cx: Scope, derived_signal: impl Fn() -> T + 'static) -> Self

Wraps a derived signal, i.e., any computation that accesses one or more reactive signals.

let (count, set_count) = create_signal(cx, 2);
let double_count = Signal::derive(cx, move || count() * 2);

// this function takes any kind of wrapped signal
fn above_3(arg: &Signal<i32>) -> bool {
    arg.get() > 3
}

assert_eq!(above_3(&count.into()), false);
assert_eq!(above_3(&double_count), true);

pub fn default(cx: Scope) -> Selfwhere T: Default,

Creates a signal that yields the default value of T when you call .get() or signal().