Trait leptos_reactive::signal_prelude::SignalWith

pub trait SignalWith<T> {
    // Required methods
    fn with<O>(&self, f: impl FnOnce(&T) -> O) -> O;
    fn try_with<O>(&self, f: impl FnOnce(&T) -> O) -> Option<O>;

    // Provided method
    fn track(&self) { ... }
}

This trait allows obtaining an immutable reference to the signal’s inner type.

Required Methods

fn with<O>(&self, f: impl FnOnce(&T) -> O) -> O

Applies a function to the current value of the signal, and subscribes the running effect to this signal.

Panics

Panics if you try to access a signal that was created in a Scope that has been disposed.

fn try_with<O>(&self, f: impl FnOnce(&T) -> O) -> Option<O>

Applies a function to the current value of the signal, and subscribes the running effect to this signal. Returns Some if the signal is valid and the function ran, otherwise returns None.

Provided Methods

fn track(&self)

Subscribes to this signal in the current reactive scope without doing anything with its value.