Trait leptos_reactive::SignalUpdate

pub trait SignalUpdate {
    type Value;

    // Required methods
    fn update(&self, f: impl FnOnce(&mut Self::Value));
    fn try_update<O>(&self, f: impl FnOnce(&mut Self::Value) -> O) -> Option<O>;
}

This trait allows updating the inner value of a signal.

Required Associated Types

type Value

The value held by the signal.

Required Methods

fn update(&self, f: impl FnOnce(&mut Self::Value))

Applies a function to the current value to mutate it in place and notifies subscribers that the signal has changed.

Note: update() does not auto-memoize, i.e., it will notify subscribers even if the value has not actually changed.

fn try_update<O>(&self, f: impl FnOnce(&mut Self::Value) -> O) -> Option<O>

Applies a function to the current value to mutate it in place and notifies subscribers that the signal has changed. Returns [Some(O)] if the signal is still valid, None otherwise.

Note: update() does not auto-memoize, i.e., it will notify subscribers even if the value has not actually changed.